Rename PyVista view file
