+++ python/pygimli/viewer/vistaview.py
